+++ src/Base/AbstractFormElement.php
@@ -29,6 +29,18 @@ abstract class AbstractFormElement implements FormElementInterface
     abstract public function getHtml(array $attributes = []): string;
 
     /**
+     * Returns the custom data.
+     *
+     * @since 2.1.0
+     *
+     * @return mixed|null The custom data or null if no custom data is set.
+     */
+    public function getCustomData()
+    {
+        return $this->customData;
+    }
+
+    /**
      * Returns the element error or null if element has no error.
      *
      * @since 1.0.0
@@ -98,6 +110,18 @@ abstract class AbstractFormElement implements FormElementInterface
     }
 
     /**
+     * Sets the custom data.
+     *
+     * @since 2.1.0
+     *
+     * @param mixed|null $customData The custom data.
+     */
+    public function setCustomData($customData): void
+    {
+        $this->customData = $customData;
+    }
+
+    /**
      * Sets whether element value is required.
      *
      * @since 1.0.0
@@ -158,6 +182,7 @@ abstract class AbstractFormElement implements FormElementInterface
         $this->error = null;
         $this->isRequired = true;
         $this->label = '';
+        $this->customData = null;
     }
 
     /**
@@ -179,4 +204,9 @@ abstract class AbstractFormElement implements FormElementInterface
      * @var string My label.
      */
     private $label;
+
+    /**
+     * @var mixed|null My custom data.
+     */
+    private $customData;
 }