+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Made RadioButtonCollection and Select countable.
 - Made RadioButtonCollection and Select iterable.
 - Added getLabel and setLabel methods to all form components.
+- Added getCustomData and setCustomData methods to all form components.
 
 ## [2.0.0] - 2018-05-14
 - Updated PHP requirement to 7.1.